backend/pattern-generator/Garment-Pattern-Estimation/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.1.8
cssselect2==0.8.0
entmax==1.0
eval_type_backport==0.3.1
filelock==3.19.1
freetype-py==2.5.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
Jinja2==3.1.6
libigl==2.6.1
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.1
networkx==3.2.1
numpy==1.23.5
packaging==26.0
pillow==11.3.0
platformdirs==4.4.0
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pycairo==1.28.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
PyYAML==6.0.3
reportlab==4.4.10
requests==2.32.5
rlPyCairo==0.4.0
scipy==1.13.1
sentry-sdk==2.54.0
smmap==5.0.2
sparsemax==0.1.9
svglib==1.6.0
svgwrite==1.4.3
sympy==1.14.0
tinycss2==1.4.0
torch==2.8.0
torch-geometric==2.6.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wandb==0.25.0
webencodings==0.5.1
WMI==1.4.9
yarl==1.22.0
fastapi
uvicorn
python-multipart
trimesh
